backend/requirements.txt
requests==2.31.0
pytest==7.4.0
requests-mock==1.11.0
pycountry==22.3.5
python-dotenv==1.2.1
